some cleanup and reorganising
