+++ enctests/testframework/utils/__init__.py
@@ -1,6 +1,9 @@
 from .utils import (
     sizeof_fmt,
+    create_clip,
     get_media_info,
+    get_source_path,
+    get_source_range,
     get_nearest_model,
     get_test_metadata_dict,
     create_media_reference,
